Ajustes no controller de pedido e mapeamento de pedido e pedidoItem.

+++ MERCEARIA/ViewModels/PedidoVM/CadastrarPedidoItemViewModel.cs
@@ -4,7 +4,7 @@ namespace MERCEARIA.ViewModels
 {
     public class CadastrarPedidoItemViewModel
     {
-        public long IdProduto{ get; set; }
+        public int IdProduto { get; set; }
         public int Quantidade { get; set; }
     }
 }
