+++ MERCEARIA/ViewModels/PedidoVM/CadastrarPedidoViewModel.cs
@@ -4,7 +4,7 @@ namespace MERCEARIA.ViewModels
 {
     public class CadastrarPedidoViewModel
     {
-        public long IdCliente{ get; set; }
+        public int IdCliente{ get; set; }
         public List<CadastrarPedidoItemViewModel> Itens { get; set; }
         public bool Pago { get; set; }
     }
